+++ src/Bridge/EasyAdmin/src/JoliMediaEasyAdminBundle.php
@@ -13,6 +13,15 @@ class JoliMediaEasyAdminBundle extends AbstractBundle
     {
         $definition->rootNode()
             ->children()
+                ->arrayNode('pagination')
+                    ->addDefaultsIfNotSet()
+                    ->children()
+                        ->integerNode('per_page')
+                            ->defaultValue(20)
+                            ->info('Number of media items to display per page.')
+                        ->end()
+                    ->end()
+                ->end()
                 ->arrayNode('upload')
                     ->addDefaultsIfNotSet()
                     ->children()
@@ -84,6 +93,7 @@ class JoliMediaEasyAdminBundle extends AbstractBundle
             ->arg('$acceptedFiles', $config['upload']['accepted_files'])
             ->arg('$maxFiles', $config['upload']['max_files'])
             ->arg('$maxFileSize', $config['upload']['max_file_size'])
+            ->arg('$paginationSize', $config['pagination']['per_page'])
         ;
     }